+++ OniAccess/ModConfig.cs
@@ -20,6 +20,7 @@ namespace OniAccess {
 		public bool FollowMovementEarcons { get; set; } = false;
 		public bool FootstepEarcons { get; set; } = true;
 		public bool ScannerDirectionEarcons { get; set; } = false;
+		public bool SweeperActivityReadout { get; set; } = true;
 
 		public float UtilityPresenceVolume { get; set; } = 1.0f;
 		public float PipeShapeVolume { get; set; } = 0.15f;